+++ tests/test_sync.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+import pytest
 
 from inspect import getmro
 from collections import OrderedDict
@@ -6,16 +7,22 @@ from collections import OrderedDict
 from mock import Mock
 from pytest import raises
 
-from pyee import EventEmitter
+from pyee import BaseEventEmitter, EventEmitter
+
 
 class PyeeTestException(Exception):
     pass
 
-def test_emit_sync():
+
+@pytest.mark.parametrize('cls', [
+    BaseEventEmitter,
+    EventEmitter
+])
+def test_emit_sync(cls):
     """Basic synchronous emission works"""
 
     call_me = Mock()
-    ee = EventEmitter()
+    ee = cls()
 
     @ee.on('event')
     def event_handler(data, **kwargs):
@@ -28,15 +35,19 @@ def test_emit_sync():
     call_me.assert_called_once()
 
 
-def test_emit_error():
+@pytest.mark.parametrize('cls', [
+    BaseEventEmitter,
+    EventEmitter
+])
+def test_emit_error(cls):
     """Errors raise with no event handler, otherwise emit on handler"""
 
     call_me = Mock()
-    ee = EventEmitter()
+    ee = cls()
 
     test_exception = PyeeTestException('lololol')
 
-    with raises(PyeeTestException) as exc_info:
+    with raises(PyeeTestException):
         ee.emit('error', test_exception)
 
     @ee.on('error')
@@ -44,7 +55,7 @@ def test_emit_error():
         call_me()
 
     # No longer raises and error instead return True indicating handled
-    assert ee.emit('error', test_exception)
+    assert ee.emit('error', test_exception) is True
     call_me.assert_called_once()
 
 
@@ -54,7 +65,7 @@ def test_emit_return():
     """
 
     call_me = Mock()
-    ee = EventEmitter()
+    ee = BaseEventEmitter()
 
     # make sure emitting without a callback returns False
     assert not ee.emit('data')
@@ -70,7 +81,7 @@ def test_new_listener_event():
     """The 'new_listener' event fires whenever a new listerner is added."""
 
     call_me = Mock()
-    ee = EventEmitter()
+    ee = BaseEventEmitter()
 
     ee.on('new_listener', call_me)
 
@@ -85,7 +96,7 @@ def test_new_listener_event():
 def test_listener_removal():
     """Removing listeners removes the correct listener from an event."""
 
-    ee = EventEmitter()
+    ee = BaseEventEmitter()
 
     # Some functions to pass to the EE
     def first():
@@ -122,7 +133,10 @@ def test_listener_removal():
     ])
 
     ee.remove_listener('event', first)
-    assert ee._events['event'] == OrderedDict([(third, third), (fourth, fourth)])
+    assert ee._events['event'] == OrderedDict([
+        (third, third),
+        (fourth, fourth)
+    ])
 
     ee.remove_all_listeners('event')
     assert ee._events['event'] == OrderedDict()
@@ -134,7 +148,7 @@ def test_listener_removal_on_emit():
     """
 
     call_me = Mock()
-    ee = EventEmitter()
+    ee = BaseEventEmitter()
 
     def should_remove():
         ee.remove_listener('remove', call_me)
@@ -149,7 +163,7 @@ def test_listener_removal_on_emit():
     call_me.reset_mock()
 
     # Also test with the listeners added in the opposite order
-    ee = EventEmitter()
+    ee = BaseEventEmitter()
     ee.on('remove', call_me)
     ee.on('remove', should_remove)
 
@@ -166,14 +180,14 @@ def test_once():
     # gets removed afterwards
 
     call_me = Mock()
-    ee = EventEmitter()
+    ee = BaseEventEmitter()
 
     def once_handler(data):
         assert data == 'emitter is emitted!'
         call_me()
 
     # Tests to make sure that after event is emitted that it's gone.
-    callback_fn = ee.once('event', once_handler)
+    ee.once('event', once_handler)
 
     ee.emit('event', 'emitter is emitted!')
 
@@ -186,7 +200,7 @@ def test_once_removal():
     """Removal of once functions works
     """
 
-    ee = EventEmitter()
+    ee = BaseEventEmitter()
 
     def once_handler(data):
         pass
@@ -204,7 +218,7 @@ def test_listeners():
     """`listeners()` returns a copied list of listeners."""
 
     call_me = Mock()
-    ee = EventEmitter()
+    ee = BaseEventEmitter()
 
     @ee.on('event')
     def event_handler():
@@ -232,7 +246,7 @@ def test_properties_preserved():
 
     call_me = Mock()
     call_me_also = Mock()
-    ee = EventEmitter()
+    ee = BaseEventEmitter()
 
     @ee.on('always')
     def always_event_handler():
@@ -262,18 +276,19 @@ def test_properties_preserved():
 
 def test_inheritance():
     """Test that inheritance is preserved from object"""
-    assert object in getmro(EventEmitter)
+    assert object in getmro(BaseEventEmitter)
 
-    class example(EventEmitter):
+    class example(BaseEventEmitter):
         def __init__(self):
             super(example, self).__init__()
 
-    assert EventEmitter in getmro(example)
+    assert BaseEventEmitter in getmro(example)
     assert object in getmro(example)
+
 
 def test_multiple_inheritance():
     """Test that inheritance is preserved along a lengthy MRO"""
-    class example(EventEmitter):
+    class example(BaseEventEmitter):
         def __init__(self):
             super(example, self).__init__()
 
@@ -289,4 +304,4 @@ def test_multiple_inheritance():
         def __init__(self):
             super(_example2, self).__init__()
 
-    a = _example2()
+    a = _example2()  # noqa